git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@4272 01821946-851b-0410-b7f4-f1477c70e57a

+++ include/newdemo.h
@@ -92,7 +92,7 @@ void NDStopPlayback (void);
 void NDStartRecording (void);
 void NDStopRecording (void);
 
-int NDGetPercentDone (void);
+float NDGetPercentDone (void);
 
 void NDRecordCreateObjectSound (int nSound, short nObject, fix maxVolume, fix maxDistance, int loop_start, int loop_end);
 void NDRecordCockpitChange (int mode);
